added footer to all pages
