+++ front-end/src/Components/Footer/Footer.js
@@ -1,0 +1,37 @@
+import * as React from "react";
+import Typography from "@mui/material/Typography";
+import Container from "@mui/material/Container";
+import Link from "@mui/material/Link";
+
+function Copyright(props) {
+  return (
+    <Typography
+      variant="body2"
+      color="text.secondary"
+      align="center"
+      {...props}
+    >
+      {"Copyright © "}
+      <Link color="inherit" href="https://exercisechallengeapp.com/">
+        Exercise Challenge App
+      </Link>{" "}
+      {new Date().getFullYear()}
+      {"."}
+    </Typography>
+  );
+}
+export default function Footer() {
+  return (
+    <Container
+      maxWidth="md"
+      component="footer"
+      sx={{
+        borderTop: (theme) => `1px solid ${theme.palette.divider}`,
+        mt: 8,
+        py: [3, 6],
+      }}
+    >
+      <Copyright sx={{ mt: 5 }} />
+    </Container>
+  );
+}
