+++ front-end/src/Pages/Welcome.js
@@ -8,24 +8,6 @@ import Container from "@mui/material/Container";
 import { Link as ReactLink } from "react-router-dom";
 import Link from "@mui/material/Link";
 
-function Copyright(props) {
-  return (
-    <Typography
-      variant="body2"
-      color="text.secondary"
-      align="center"
-      {...props}
-    >
-      {"Copyright © "}
-      <Link color="inherit" href="https://exercisechallengeapp.com/">
-        Exercise Challenge App
-      </Link>{" "}
-      {new Date().getFullYear()}
-      {"."}
-    </Typography>
-  );
-}
-
 function Content() {
   return (
     <React.Fragment>
@@ -76,20 +58,6 @@ function Content() {
         </ReactLink>
       </Container>
       {/* End hero unit */}
-
-      {/* Footer */}
-      <Container
-        maxWidth="md"
-        component="footer"
-        sx={{
-          borderTop: (theme) => `1px solid ${theme.palette.divider}`,
-          mt: 8,
-          py: [3, 6],
-        }}
-      >
-        <Copyright sx={{ mt: 5 }} />
-      </Container>
-      {/* End footer */}
     </React.Fragment>
   );
 }
